refactor(gateway): split channel-binding tests out of session/tests.rs
The new unbind tests pushed session/tests.rs to 1550 lines, past the 1500-line
CI limit. Move them to a sibling test module alongside archive_search_tests,
following the same #[path] wiring.
